backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
email-validator==2.3.0
python-multipart==0.0.20
PyYAML==6.0.2
